% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta-01-getdata.R
\name{clean_fasta_data}
\alias{clean_fasta_data}
\title{Clean response data from FaSta API}
\usage{
clean_fasta_data(x)
}
\arguments{
\item{x}{Return object from \code{retrieve_fasta_data}.}
}
\description{
Clean response data from FaSta API
}
\seealso{
Other FaSta: \code{\link{create_fasta_log}},
  \code{\link{retrieve_fasta_data}}
}
\concept{FaSta}
