% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta-01-getdata.R
\name{create_fasta_log}
\alias{create_fasta_log}
\title{Create log data from FaSta API response}
\usage{
create_fasta_log(x)
}
\arguments{
\item{x}{Return object from \code{retrieve_fasta_data}.}
}
\description{
Create log data from FaSta API response
}
\seealso{
Other FaSta: \code{\link{clean_fasta_data}},
  \code{\link{retrieve_fasta_data}}
}
\concept{FaSta}
