% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp.opendata-package.R
\docType{package}
\name{dp.opendata-package}
\alias{dp.opendata-package}
\title{\pkg{dp.opendata} - Query DB Opendata APIs}
\description{
\pkg{dp.opendata} - Query DB Opendata APIs
}
\keyword{DB}
\keyword{package}
