% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta-01-getdata.R
\name{keep_state_changes}
\alias{keep_state_changes}
\title{Keep data poits where state changes occur}
\usage{
keep_state_changes(dt, change_index = FALSE)
}
\arguments{
\item{dt}{dataframe from \code{clean_fasta_data}.}
}
\description{
Keep data poits where state changes occur
}
