% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta-01-getdata.R
\name{retrieve_fasta_data}
\alias{retrieve_fasta_data}
\title{Query FaSta API}
\usage{
retrieve_fasta_data(path = "fasta/v2/facilities", key = NA)
}
\arguments{
\item{path}{Path to FaSta API.}
}
\description{
Query FaSta API
}
\seealso{
Other FaSta: \code{\link{clean_fasta_data}},
  \code{\link{create_fasta_log}}
}
\concept{FaSta}
